Move things around
- teamloader tests were testing config stuff
- the secrets package didn't do anything with the secrets
- add functional args to the teamloader, we now only need the "Load"
  function
