+++ pkg/oci/package.go
@@ -13,8 +13,8 @@ import (
 	"github.com/google/go-containerregistry/pkg/v1/static"
 	"github.com/google/go-containerregistry/pkg/v1/types"
 
-	"github.com/docker/cagent/pkg/config"
 	"github.com/docker/cagent/pkg/content"
+	"github.com/docker/cagent/pkg/path"
 )
 
 // PackageFileAsOCIToStore creates an OCI artifact from a file and stores it in the content store
@@ -24,7 +24,7 @@ func PackageFileAsOCIToStore(filePath, artifactRef string, store *content.Store)
 	}
 
 	// Validate the file path to prevent directory traversal attacks
-	validatedPath, err := config.ValidatePathInDirectory(filePath, "")
+	validatedPath, err := path.ValidatePathInDirectory(filePath, "")
 	if err != nil {
 		return "", fmt.Errorf("invalid file path: %w", err)
 	}
