+++ pkg/teamloader/testdata/basic.yaml
@@ -1,0 +1,6 @@
+version: "2"
+
+agents:
+  root:
+    model: openai/gpt-4o
+    instruction: Be good
